refactored the base dao classes into a single rootbasedao to cut back on replicated code, while still maintaining the three strongly typed base dao options
Refactor data access classes to use RootBaseDao

This commit refactors the `BaseDao`, `BaseGetAllDao`, and `BasePagedDao` classes in the `Guid`, `Int`, and `Long` namespaces to inherit from a new abstract class `RootBaseDao<T, TIdType>`. This change centralizes common functionality, reduces code duplication, and streamlines the data access layer. The `Get` method in `BaseDao` has been overridden to utilize the `Dao` property for fetching entities. The new `RootBaseDao` class includes methods for CRUD operations and transaction management.
